requirements.txt
absl-py==0.11.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
astunparse==1.6.3
attrs==20.3.0
cachetools==4.2.0
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
chex==0.0.3
cycler==0.10.0
distlib==0.3.1
dm-tree==0.1.5
filelock==3.0.12
flatbuffers==1.12
flax==0.3.0
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hydra-core==1.1.0.dev1
identify==1.5.11
idna==2.10
importlib-resources==4.1.1
iniconfig==1.1.1
isort==5.7.0
jax==0.2.7
jaxlib==0.1.57
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
msgpack==1.0.2
nodeenv==1.5.0
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.1.0.dev14
opt-einsum==3.3.0
optax==0.0.2
packaging==20.8
Pillow==8.1.0
pluggy==0.13.1
pre-commit==2.9.3
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.6.0
shapecheck==0.0.2
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
toml==0.10.2
toolz==0.11.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
